refactor: group actions doc editor
