+++ studio/src/app/components/editor/doc/app-add-paragraph/app-add-paragraph.scss
@@ -1,6 +1,6 @@
 @use "../../../../../global/theme/mixins/button";
 
-app-actions-doc-editor {
+app-add-paragraph {
   display: block;
 
   position: absolute;
